chore: add unit test runner
This adds Jest as a unit test runner. We have it configured to run any
test files within the packages, which it will do in parallel. This means
our tests should remain nice and fast regardless of how many packages we
add to the repo.

I added an example test just to illustrate where they'll be written.
The tests should also automatically run in this PR.