chore: add automated type checking
We now explicitly check types with TypeScript, so that if we provide
valid JSDoc comments for our code we can be sure that everything is
type-safe. This will hopefully result in fewer unexpected bugs and
increase confidence in our tooling!

This also gives us useful hinting in VS Code, so hopefully writing the
code will be easier for us as well.