chore: use Babel with ESLint
ESLint does not support many ES2020 features, e.g. class public and
private fields. It'd be nice to be able to use these considering that
we're only targeting environments where they're available.

This addresses the issue and makes sure that both ESLint and TypeScript
do not complain when we use modern features.