feat: initial bootstrapping of the errors package
This is a starting point for some central errors. Most of the code in
here is really to act as an example and it will probably change a lot in
future. This PR is mostly here for us to test that the release process
is working as expected.

This will allow me to delete the "example" package (in a separate PR).