feat: add a method to check for operational errors
This method is mostly being added so that I can legitimately bump the
package version (see the commit footer which indicates a version).

I think this is still a useful method anyway, as ideally an engineer
should be able to choose to decorate existing errors with an
`isOperational` property if it's not possible to use our classes for
some reason.

Release-As: 0.1.0