chore: write a script to create new packages
When I was setting up `packages/errors`, I missed a lot of things that
were required to get the package to be released with Release Please. I
also think it'll be easy to see some inconsistency in the way packages
are built the more we add them.

This PR adds a new script which automatically creates the boilerplate
for a new Reliability Kit package, taking a lot of inspiration from the
equivalent script in Tool Kit.

There's documentation on how the script is run in the contributing
guide, as I think that's where people would most likely look.