feat: add a serialize-error package
This adds a serialize-error utility package. This new package exports
a utility function to serialize an error object in a way that's friendly
to loggers, view engines, and converting to JSON.

I'm writing this because we need a reliable way to serialize error data
if we tackle any of the following:

  * An error logging middleware so that thrown errors are logged
    consistently to Splunk (planned)

  * Modifications to `n-logger` to enforce more consistent logging of
    errors (not planned but likely we will need it)

  * A debugging page middleware which renders an HTML error page to help
    engineers debug while working locally (planned)

I think it makes sense to publish this as a separate package which is a
dependency of these others.

Release-As: 0.1.0