fix: change the default status code to null
After chatting in the PR, I agree with Alex that not all errors should
have a status code. There are contexts outside of HTTP (e.g. a worker
process or queue processor) in which an HTTP status makes no sense. I
think setting a null value makes sense as it indicates the error is
explicitly not HTTP-related.