feat: add a log-error package
This splits out the main logging functionality from the middleware into
a new package which allows us to log handled, recoverable, and unhandled
logs consistently no matter where we need them (not just in middleware).
This means we can support more non-Express uses-cases and will massively
simplify the code in the Express middleware.

Release-As: 0.1.0