feat: add new app properties to error logs
This adds new app-related properties to error logs. We decided these
would be useful to help us debug issues and will bring logs in Splunk
closer to Sentry in terms of the information they hold.

These are the added properties:

  - `app.commit`: set to `process.env.HEROKU_SLUG_COMMIT`
  - `app.nodeVersion`: set to `process.versions.node`
  - `app.releaseDate`: set to `process.env.HEROKU_RELEASE_CREATED_AT`

Some of these rely on the Heroku Dyno Metadata labs feature but we
recommend use of this in all our apps anyway:
https://devcenter.heroku.com/articles/dyno-metadata

**Note to reviewers:** most of the changes here are to the tests, which
all need to be updated when we add new properties to fields. I think the
tests themselves need less of a review as it's a lot of repetition.

This resolves FTDCS-248.