feat: add a middleware-render-error-info package
This makes a start on FTDCS-184, adding a new middleware function which
renders an error information page for debugging. This is the absolute
bare-bones of the module so far and it only provides some basic error
information.

There are also some unanswered questions, mostly whether this will play
nicely with the n-raven and n-express error handlers. I think this will
need some testing once we have a basic 0.1.0 release of this package.

This does not include documentation yet, I think I'd rather work on that
once we've nailed down exactly what the behaviour is and whether we
include any config options.

Release-As: 0.1.0
Co-Authored-By: Alex Muller <alex.muller@ft.com>
Co-Authored-By: Andy Gout <andy.gout@ft.com>