chore: remove the Express example tests
It doesn't really make sense to be testing in the Express example app,
as it confuses the meaning of that folder. Instead we should be running
integration tests in each package with a potential top-level package
later which tests how they behave together.

I'm confident deleting this because actually all we were testing was:

  * The serialize-error and serialize-request packages, which already
    have good tests and 100% coverage. Their unit tests are enough

  * That our mock JSON error render was working as expected, and this
    isn't something we're shipping or expecting users to add

All-in-all this doesn't change the test coverage for anything that
matters, and it allows us to focus the examples folder on giving
Reliability Kit users some examples without any testing stuff confusing
things.