docs: clarify that x-request-id is auto-logged
We weren't being very clear that the `X-Request-ID` header is already
automatically logged under the `request.id` property. I've attempted to
clarify this in the documentation.

If this is still not clear then another option would be to ignore this
header if it's passed in as part of the `includeHeaders` option.