fix: refactor to centralise environment variables
This tucks away nearly all of our environment variable access into a new
module named app-info. This means we're no longer repeating some logic
across the other modules and can remove a lot of tests.

This also adds a feature: we're now not only looking at the
`SYSTEM_CODE` environment variable, we're also attempting to load the
`package.json` file in the current working directory and normalizing the
name to get a default system code. This nearly* matches the behaviour in
n-express.

A benefit of this work is that if we need to use the system code in
other places (which I think is likely, e.g. if we wrap any of the
tooling for OpenTelemetry). We could also replace usage of all these
environment variables across our other libraries and apps so that
there's a single source of truth for getting the app's system code.

---

*The system code sanitization only removes a prefix of `ft-` which
doesn't quite match the n-express behaviour. There are two things in
this that I don't think we need any more
https://github.com/Financial-Times/n-express/blob/main/src/lib/normalize-name.js

  1. The n-express library is designed to normalise a _name_ and not a
     system code, so it removes the `next-` prefix as well. We don't
     want to do this as many of our system codes _do_ begin with `next-`

  2. The n-express library removes a suffix of `-vXXX` (`X` being any
     number). From what I can tell, at some point we intended to version
     package names but this is not used in any of our current apps.
     https://github.com/Financial-Times/n-express/commit/a598c601040aabbc35e06efc1cd0713cf692d1b5