docs: clarify our decision to write CommonJS
I saw this decision document on API Rationalisation and realised that
maybe we hadn't clarified our choice of CommonJS. I can see it coming up
in future so I decided to write up my original thoughts.

https://financialtimes.atlassian.net/l/cp/cUUXy7Gx

This can change in future, but think it's worth documenting why things
are the way they are!