fix: make sure that ESM + TypeScript works
We messed up here, and we didn't consider the use case of TypeScript
written as ES Modules being compiled to CommonJS modules. When using the
import signature:

```
import example from 'example';
```

the JavaScript compiled from the above TypeScript will always look for
the `default` property on the module. This is different from Node.js
native ESM which is what we tested with.

The solution is to alias `module.exports` as `module.exports.default`
which ensures that this works in the same way for native ESM as well as
compiled JavaScript from TypeScript.