chore: add commands to generate Type declarations
We can now generate TypeScript type declaration files for all packages
by running `npm run build`. This ensures that TypeScript projects which
import Reliability Kit modules will get proper type hinting and will not
throw errors during compliation.

This is a known issue, and it'd be nice if we didn't have to do it, but
it's either this or go wholesale TypeScript which involves a more
complex build step and it'd be a large-scale migration. I also still
think we shouldn't do it.

More context on the issue:

  * https://github.com/microsoft/TypeScript/issues/19145
  * https://github.com/microsoft/TypeScript/issues/29824