feat: add an uncaught exception handler
This is the basic functionality for an uncaught exception handler. We're
going with a single option for now so that you can pass the process in
(e.g. for binding to a child process).

I decided against adding an `exitCode` or `onError` option as outlined
in the issue, because I think we shouldn't add things unless our teams
have demostrated a need for them.

Resolves #225.

Release-As: 0.1.0