chore: configure Dependabot to use fix commits
This configures Dependabot to use the `fix:` commit prefix for
production dependencies by default and `chore:` for development
dependencies. Right now it's defaulting to `chore:` for everything
which is incorrect as chores do not trigger releases.

This stops us from having to manually override the commit message so
often, though we'll still need to do this if we decide that the
dependency bump is either a feature or breaking change.