feat: add the ability to transform log data
This adds an option to specify log transforms - functions which can
modify log data before it's sent. This is a foundational piece of work
which is needed in order for us to provide feature parity with
n-mask-logger.

A later piece of work will introduce a log transform which masks fields
in exactly the same way as n-mask-logger.