feat: add a mask transform
This should be the final piece of work needed to allow users of
n-mask-logger to migrate to this library in future. It brings us to the
point where we have a functional replacement for all three of our
current logging libraries.

There's more information in the README and migration guide for this
commit, but essentially you can now introduce the same masking-behaviour
as n-mask-logger to the Reliability Kit logger:

```js
const { Logger, transforms } = require('@dotcom-reliability-kit/logger');

const logger = new Logger({
    transforms: [
        transforms.legacyMask()
    ]
});

logger.info({
    email: 'oops@ft.com'
});
// Outputs:
// {
//     "level": "info",
//     "email": "*****"
// }
```

I went with the name `legacyMask` so that we can continue to support
this masking behaviour alongside a new one if we ever decide to improve
it - I noticed quite a few issues while digging into the n-mask-logger
source code and it's not as thorough as we might like it to be.