feat: add prettification to logs in local dev
This adds the ability for an app using the new logger to see prettified
logs locally, which makes issues easier to debug. We configure
pino-pretty if it's present to work with our logs.

This resolves #281. I decided against adding extra options to disable
prettification, as teams can manually set their `NODE_ENV` to
`production` if they want to inspect the raw JSON output.

The issue linked above contains more information on why I decided to
solve this within the logger rather than externally or with Tool Kit.