fix: stop pinning dependencies in express example
This is causing issues because it results in multiple versions of
Reliability Kit packages being installed in the repo. This means that,
when we're testing changes to the types, TypeScript gets confused about
which version of a package we're actually looking at. Setting these to
`*` means that we always use the local dependency.

We also set bumping dependencies in the examples to always be marked as
a chore, as this should never result in an actual release.