feat: add the ability to set a custom logger
Setting a custom logger is essential for teams who use log masking or
their app relies on other custom logging (e.g. their logger must include
certain base data).

This is more complex than I originally thought, as we can no longer
trust the logging methods that we're using. Because of this I'm now
wrapping all our error logging in a try/catch so that we can fall back
to logging via `console.log` which is guaranteed to work. This is a
similar approach to @reliability-kit/logger.

This resolves #221.