docs: add a serverless (Lambda) example
This example shows how Reliability Kit can be used in a Serverless
framework app which is deployed to AWS Lambda.

It is not deployable in its current state because of the way the
monorepo manages package-lock.json and node_modules. In order for it
to be deployable, we would need to make a directory which contains
a node_modules/ with every dependency (and its tree) that the Lambda
requires.