docs: move the example apps to separate repos
These were causing a few issues while they remained a part of the main
monorepo:

1. Express and Serverless are large complex modules. Installing them as
part of builds and local development slows things down unnecessarily.

2. Because Express is installed as a production dependency in the
example, Dependabot gets confused and always opens express version
bumps as "fix" level commits. This ends up creating noisy and
unnecessary releases for packages which install Express as a development
dependency (we do this in order to run end-to-end tests)

The new examples (mostly the same but with a few added comments) live in
these two new repos:

  * https://github.com/Financial-Times/reliability-kit-example-express
  * https://github.com/Financial-Times/reliability-kit-example-serverless