feat: add filter option to the logging middleware
This allows an app to tell the error logging middleware not to log for
specific errors or requests. I think this is required for us to use the
error logging middleware in next-api because it filters out certain
logs based on which app has made the request.

Without a filter for these errors we'll see a dramatic noisy increase in
next-api logs.

I don't want this option to start being used frequently across our apps
so there's a warning in the README. I'm also sending further logs if
someone misconfigures the option so that we have visibility if an app
has broken logging due to use of this feature.