feat: support Node.js 18 error signatures
In Node.js 18, the constructor signatures for errors have changed to
support setting both a message and options:
https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Error/Error

Our custom errors did not support this new constructor signatures, this
work ensures that they work in the same way as Node.js 18 native errors.
See issue #438 for more information.

We now support the following signatures:

```diff
  new OperationalError();
  new OperationalError(message);
  new OperationalError(options);
+ new OperationalError(message, options);
```

Caveat: as part of this work, I found that a limitation of TypeScript +
JSDoc is that you can't document multiple different overloads for the
same method. In TypeScript you can:

https://www.typescriptlang.org/docs/handbook/2/functions.html#function-overloads

This means our types aren't _optimal_ but I think it's fine for now. In
TypeScript v5, overloads will be supported in JSDoc:

https://devblogs.microsoft.com/typescript/announcing-typescript-5-0-beta/#overload-support-in-jsdoc

This PR resolves #438.