feat: add a crashing errors dashboard
This adds a dashboard to get visibility of crashing errors across all
apps which use Reliability Kit. It's similar to the Heroku dashboard
that we already have but with a focus on crashes (`UNHANDLED_ERROR`
events) across all apps.

You can apply a filter if you need to limit to a specific team or group.