fix: add eslint as a peer dependency
This adds eslint as a peer dependency, because otherwise we get some
incompatibilities when installing within a repo. I just tested in
dotcom-debug-user and got some cryptic errors because it wasn't using
the latest version of ESLint.

I didn't consider this when reviewing the original PR :)

By adding a peerDependency of `eslint@8.27.0` we ensure that all the
rules we specify actually exist within the app's installed version of
ESLint. The one that immediately threw an error in dotcom-debug-user
was https://eslint.org/docs/latest/rules/no-new-native-nonconstructor

Peer dependencies are designed for this case. They're meant to be used
when a module _requires_ that it's installed alongside a specific
version of another module. There's a good blog post on the subject here
if not familiar: https://nodejs.org/en/blog/npm/peer-dependencies