fix: check for supported pino-pretty versions
I'm gonna start by saying that I hate this. This is a first attempt at
fixing issue #516 which is caused by the app installing a version of
pino-pretty older than v10. This can easily happen unfortunately.

What I hate about this solution is how complex it is, but I'm not sure
if there's a nice alternative. What this solution does is loads in the
pino-pretty `package.json` file, checks the version, and ensures that it
meets our minimum version requirements. If this doesn't work then we log
a warning and don't try to use the prettifier. It's just so complex!

Some alternatives I've considered are:

1. Get rid of the pino-pretty integration altogether and make log
prettification a problem for the app which uses the Reliability Kit
logger. This has the added benefit of solving #474. It introduces a new
issue that any apps which prettify logs will have to add configuration
to expect a `message` rather than `msg` (pino's gross default) property.
We'd also have to release a major version to drop the prettifier.

2. Define pino-pretty v10 as a `peerDependency` with an
`peerDependencyMeta` of `optional: true`. This has the benefit of not
installing pino-pretty alongside the Reliability Kit logger but there is
a warning output when you install an incompatible version of
pino-pretty. The issue is that people mostly ignore npm install warnings
so I think we'll still get support requests.

Thoughts? I think alternative 2 might be the most sensible approach with
alternative 1 as a further future change? Ugh.