fix: introduce pino-pretty as a peer dependency
This fixes an issue we have with older versions of pino-pretty being
installed by consuming applications. We get cryptic errors in the code
(see #516) rather than a warning when we run `npm install`.

Adding pino-pretty as an optional peer dependency resolves this issue.
This range will have to expand as new versions of pino-pretty are
released, so we need to be wary of Dependabot PRs relating to it. I'll
see if I can add a specific rule to our Dependabot config as a separate
piece of work.