fix: throw an error when log level is not a string
If `log()` is called with a non-string first argument then we can't
actually do anything useful with it - it's not valid based on the
TypeScript types and so we should also error when we encounter this
issue at runtime.

I'm throwing outside of the catch block because this is a rare occasion
where we _do_ want logging to fully fail and an error to be thrown as
it's a clear programmer error - the only resolution is to change the
code to use a string.

Resolves #547.