feat!: switch to Reliability Kit logger
This removes n-logger from all packages, switching it to Reliability Kit
logger. To make sure that teams can undo this temporarily, I've added
the ability to configure the logger to the only module where it wasn't
possible - middleware-render-error-info.

This means that these modules will no longer log to Splunk unless
the app they're used in is configured to forward `stdout` logs there.