feat!: log unhandled errors with a level of "fatal"
Now that we've moved to Reliability Kit logger we can log unhandled
errors with a level of "fatal". If the logger doesn't have a fatal
method then we fall back to the error method so this can continue to
work with the old loggers.

This is one step towards us using appropriate log levels in all cases
across our estate and being able to more easily differentiate crashing
bugs without having to know the event name.