feat: add a BaseError class
Reliability Kit provides a neat way to set additional properties on
error objects during construction, but all the errors we provide are
marked as operational. This means that Reliability Kit errors ideally
shouldn't be used in library code and we lose the convenience of our
extended constructors.

This PR adds in a `BaseError` class which is marked as _not_
operational. The `OperationalError` class now extends this so we're not
duplicating logic.

This resolves #238.

See https://github.com/Financial-Times/cp-content-pipeline/pull/423 for
a real world use-case.