fix: correct types for the logger
The types that were generated weren't working properly in TypeScript
projects that imported and created a new `Logger` instance, this is
because we were assigning directly to `module.exports` instead of
creating a new instance of the logger and _then_ assigning it.

I generated the new type files and tested them in a TS project
(n-user-api-client), they now work as expected.