feat: handle common errors thrown by fetch
This adds in handling for most of the common errors that we see when
making `fetch` requests. There's a lot of extra logic in here but I
think it's worthwhile to augment these errors and throw ones with
correct status codes that our engineers can immediately understand.

Due to the fact that we use a variety of different `fetch`
implementations across the FT, this really isn't doable at the Splunk
level: we can't reliably say "here's a search which shows every fetch
request that has timed out.

Because of the amount of logic and the fact we're relying on error names
and codes, this comes with a suite of end-to-end tests that allow us to
verify that the following `fetch` implementations throw the expected
errors. This way, if something changes, we'll see it immediately in a
failing dependabot PR:

  - Node.js 18+ native fetch (powered by Undici)
  - Manual [Undici](https://github.com/nodejs/undici) installs
  - [node-fetch](https://github.com/node-fetch/node-fetch) v1.x
  - [node-fetch](https://github.com/node-fetch/node-fetch) v2.x