fix: trim long request URLs
When the request URL is particularly long, Splunk events break because
they're concatenated if they exceed a certain length. This causes the
JSON to be invalid and so we don't parse out fields.

We probably shouldn't blindly trust the URL as it's user input. We now
truncate request URLs at 200 characters and add an indicator so that
it's clear in the logs when the URL was too long.