feat: expose error fingerprints as an HTTP header
This sets a new header in the error rendering middleware which exposes
the fingerprint of an error if it has one. This allows us to better
debug in production environments - our logs also include this
fingerprint so that you can map our generic error pages to the type of
error that occurred and where in the code it was thrown.

Some considerations:

  * If an app is behind a CDN and error pages are cached, then we will
    need to vary on the `x-error-fingerprint` header in order for this
    to be useful - otherwise the first fingerprint will get cached

  * Not all apps will get this functionality at the same time. They'll
    need to either start using the error rendering middleware or bump
    the major version, because this is landing as part of a breaking
    change