chore!: always log ISO datetime
This removes the ability to configure the time property of the output
logs, setting it to an ISO 8601 formatted string with the current date
and time. This is a breaking change because the logs will be different
and any tooling that relies on these may break.

There's more detail about why we made this choice in #805.

Resolves #805.