fix: switch back to using a pino-pretty transport
Back in #802 we fixed an issue where we were doubling up logs due to
pino-pretty being installed. We switched to streaming logs into
pino-pretty instead of configuring it as a transport.

This turned out to be a configuration issue which allows us to switch
back to configuring pino-pretty as a transport rather than creating a
stream. I prefer this approach as it's easier to test and I think
results in more readable code.

This is largely a revert of 7abae5a92a012b835a009e4df983bc401223e134,
but some related code has changed since this commit so I had to manually
switch back to the old way of doing things.