fix: add manual types for serialize-request
This is an experiment to see how much more/less code we'd have to
maintain if we write manual types for our modules.

The con of this work is that we have to maintain some TypeScript
alongside our source JavaScript. The pros are that we don't have to
write as much JSDoc and also we can fully remove the build step.