chore: stop generating type declaration files
Because our code is CommonJS, the generated type declaration files just
don't really work in an ESM context at all. It results in a lot of weird
workarounds. We've decided to start writing our type declarations
manually so that we have more control and can offer better types to
users of Reliability Kit.