fix: handle node-fetch memory leak
node-fetch has a memory leak if you don't read the body of a response.
Within the library we don't ever know which fetch implementaton we're
using so we've had to test for a property (`pipe`) that only exists on
the node-fetch response body and not native fetch.

If we're throwing an error and we see that the response body has this
method, then we pipe the body into a black hole stream. This means that
the body is unusable if there was an error but we can document this
limitation. I'd rather not mess with trying to read/parse data from the
response body if possible as it adds a lot of complexity and I don't
want to add _too_ many cases to deal with node-fetch when we'll be
moving to native-fetch relatively soon.

See-Also: https://financialtimes.atlassian.net/browse/CPREL-1047