chore!: stop using default exports
This is going to cause us lots of problems when we want to export new
methods, e.g. methods to send custom metrics which I'll be doing in a
later commit. Nobody is using the JavaScript API at the moment and so
this is a safe major version bump for most. It's also worth addressing
this while use is still low.

See #945.