fix: wrap the Reliability Kit logger
We have been losing information from the OpenTelemetry internal logs
because they log multiple messages in a row which pino/reliability kit
does not support. We don't support this because we want to maintain
consistency with n-logger.

To fix this, we need to wrap the Reliability Kit logger and extract any
extra logged information into a new property. I've gone with `details`.
Testing this locally, it's not common to get any info but it's important
when we do get it - it helped me debug the issue with loading Pino ahead
of instrumentation.