fix: log a better warning if log level is invalid
Previously we were logging that the level was deprecated if we didn't
recognise it. This made it difficult for Content Discovery to debug a
logging issue. Now the same issue will log a more specific warning
alongside an error so you can trace it back to the place in the source
code where the level is being used.