fix: instrument node-fetch properly
Due to an issue with the Node.js http auto-instrumentations, requests
made with node-fetch do not send metrics. See the code comment for the
relevant issue.

Requiring `https` here fixes the issue. I think it's best if we solve
this in our client rather than require our teams to update their apps.
Technically removing this hack will be a breaking change - I think
that's fine, we can coordinate with the removal of node-fetch from our
libraries and the migration to native fetch.