feat: manage bad requests with 'allow request methods' package
Current behaviour: a DDoS can use an unexpected HTTP method to bypass the Fastly cache and cause
interruption to service. For example, making a request to the homepage as a POST rather than a GET.
Express will return a 404 for this error.

Ideal future work: If we can transform this 404 error into a 405 Method Not Allowed, we may be able
to use Signal Sciences to block this kind of attack very quickly. This is because it’s highly
unlikely that a genuine user will accidentally make a POST request.

Definition of done for this ticket: Build and release a package containing an Express middleware
that applications can consume which will cause unexpected request methods to error with a 405 rather
than a 404.

See Also: [CPREL-1276]