feat: disable prettifier for more prod env names
This disables log prettification if the environment is either `prod` or
`p`, which we use across some systems. This is done in the logger rather
than app-info because that would be a breaking change. If app-info
normalises the environment in a later version then we can switch logger
back to expecting `production` only.

Co-Authored-By: Matt Hinchliffe <i-like-robots@users.noreply.github.com>
See-also: #1368